[ fix ] 서버리스 함수 12개 제한으로 인한 api 통합
