--- api/cakes/getPgTokenData.ts
@@ -1,8 +1,0 @@
-import PATH from '../common/path';
-import { client } from '../common/axios';
-
-export const getPgTokenData = async (pgToken: string | string[] | undefined) => {
-  const data = await client.get(`${PATH.CAKES}/${PATH.APPROVE}?${PATH.PG_TOKEN}=${pgToken}`);
-
-  return data;
-};
