+++ api/letters/getCakeLetters.ts
@@ -1,18 +1,20 @@
 import { client } from '@/api/common/axios';
-import PATH from '@/api/common/path';
+import PATH from '@/constant/path';
 
-export const getCakesLetters = async (wishId: string | string[] | undefined, cakeId: string | string[] | undefined) => {
+export const getCakesLetters = async (
+  wishId: string | string[] | undefined,
+  cakeId: string | string[] | undefined,
+) => {
   const accessToken = localStorage.getItem('accessToken');
 
   const url = `${PATH.API}/${PATH.V1}/${PATH.CAKES}/${wishId}/${cakeId}`;
 
-  const data = await client.get(url,
-    {
-      headers: {
-        'Content-Type': 'application/json',
-        Authorization: `Bearer ${accessToken}`,
-      },
-    });
+  const data = await client.get(url, {
+    headers: {
+      'Content-Type': 'application/json',
+      Authorization: `Bearer ${accessToken}`,
+    },
+  });
 
   return data.data.data;
 };
