+++ api/letters/getCakesCount.ts
@@ -1,5 +1,5 @@
 import { client } from '@/api/common/axios';
-import PATH from '@/api/common/path';
+import PATH from '@/constant/path';
 
 export const getCakesCount = async (wishId: string | string[] | undefined) => {
   const accessToken = localStorage.getItem('accessToken');
@@ -11,7 +11,6 @@ export const getCakesCount = async (wishId: string | string[] | undefined) => {
       'Content-Type': 'application/json',
       Authorization: `Bearer ${accessToken}`,
     },
-
   });
   return data.data.data;
 };
