+++ api/login/sendCodeToServer.ts
@@ -1,10 +1,12 @@
 import { client } from '@/api/common/axios';
-import PATH from '../common/path';
+import PATH from '../../constant/path';
 
 export const sendCodeToServer = async (authCode: string) => {
   const url = `${PATH.API}/${PATH.V1}/${PATH.AUTH}/${PATH.KAKAO}/${PATH.CALLBACK}`;
 
-  const data = await client.post(url, {},
+  const data = await client.post(
+    url,
+    {},
     {
       headers: {
         'Content-Type': 'application/json',
@@ -12,7 +14,8 @@ export const sendCodeToServer = async (authCode: string) => {
       params: {
         code: authCode,
       },
-    });
+    },
+  );
 
   return data.data.data;
 };
