+++ constant/path.ts
